requirements.txt
torch
opencv-python
numpy
scipy
pillow
argparse
torchvision

-e .